feat: 更新部分をrepositoryに移譲
